arithmetic Funtions App/requirements.txt
numpy==2.2.3
Pillow==11.1.0
reportlab==4.3.1
streamlit==1.42.1
